RavenDB-4008: Unified the PrefixTreePageHeader derived structs to extend the original PageHeader.
Changing the layout to avoid having to offset and unify with usage in other structures also allowed to have an smaller Table and Page header because we could use original PageHeader padding to actually store data. Also we are being more explicit about the size of structs adding the Padding property and Size to the Explicit Layout definition.