RavenDB-4008: Direct Pointers to Disk and Free space overhaul.
After a review done at QCon we decided that the best thing was to avoid virtual pointers altogether in favor of direct disk locations. The rationale is: "we wont move those pointers anyways, so why pay the cost of keeping the virtual tables around."
We also restructured how the free space is being handled, before we were using a bitmap, the problem is that keeping the bitmap updated causes IO we would not want to have; therefore we are migrating (as a first step) into a single fixed size tree.