RavenDB-4685: SliceComparer.InRange optimization
Took 3 whole days of tinkering but, if I would have travel from the future and told myself that I would have been able to optimize at this level, I wouldnt had believed myself. The SliceComparer.InRange is faster than doing 2 independent comparisons even though the code is a mess of "if" statements (which is BAAAD in micro-architectural terms). I even had to slap a warning in there to avoid anyone to break it.
The RecentlyFoundPages cache (which is the only one that uses this method so far is 37% faster than the original.

The assembler code generated is awesome, the JIT did a great job there.