RavenDB-4762: Non-Complete Voron implementation of the Cedar (Efficiently-updatable double-array trie)
Cedar implements an updatable double-array trie, which offers fast update/lookup for skewed queries in real-world data, e.g., counting words in text or mining conjunctive features in a classifier. Its lookup speed is of the the C++ implementation is significantly faster than C implementations of modern cache-conscious tries; even when keys are uniformly, randomly queried, its update/lookup speed is comparable to std::unordered_map.

The Voron version also implemented a few new operations to support many of the BTREES operations like:
 - Successor
 - SuccessorOrEqual
 - Predecessor
 - PredecessorOrEqual
 - Page Splitting
 - Tree Iterators

Optimizations specific for a Paged System like Voron:

* Avoid constructing an struct on the stack and then copying it into memory.
* Improved assembler for the pointer arithmetics using multiplicative modulus.
* Avoid prematurely mark as write a page where only data is being read.
* Improvements using constants instead to avoid idiv operations.
* Reference passing
* Faster dedicated PageLocator.
* All pages are treated individually.
* Splitting optimization using removal of keys instead of creation of a whole new tree.
* Separated and simplified access to the pointers using struct based memory assessors.
* Not allocating nodes until they are going to be used.
* Space limited nodes with custom layouts.
* CPU and Memory allocation optimizations

This version is based on the ported version of the original code which is licensed under BSD.
http://www.tkl.iis.u-tokyo.ac.jp/~ynaga/cedar/